ai-backend-python/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
openai>=1.3.5
pymilvus>=2.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
pdfplumber>=0.10.3
python-docx>=1.1.0
redis>=5.0.1
httpx>=0.25.2
pydantic>=2.5.2
pydantic-settings>=2.1.0
tenacity>=8.2.3
